has_warning {testit}	R Documentation

Check if an R expression produces warnings or errors

Description

The two functions has_warning() and has_error() check if an expression produces warnings and errors, respectively.

Usage


has_warning(expr)

has_error(expr, silent = !interactive())

Arguments

expr
an R expression

silent
logical: should the report of error messages be suppressed?

Value

A logical value.

Examples


has_warning(1 + 1)
has_warning(1:2 + 1:3)

has_error(2 - 3)
has_error(1 + "a")
has_error(stop("err"), silent = TRUE)

[Package testit version 0.13.1 Index]